pyproject.toml
[project]
name = "sd-style-lab-images-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
